rename DetailViewSidebarUpsell -> JsonViewerSidebarUpsell
