+++ apps/studio/src/components/upsell/JsonViewerSidebarUpsell.vue
